Chapter10/chatbox/requirements.txt
fastapi
uvicorn

python-dotenv
openai
